+++ URLSession/NetURLSessionDelegate.swift
@@ -51,6 +51,13 @@ extension NetURLSessionDelegate: URLSessionTaskDelegate {
         completionHandler(.continueLoading, nil)
     }
 
+    @available(iOS 11.0, tvOS 11.0, watchOS 4.0, OSX 10.13, *)
+    func urlSession(_ session: URLSession, taskIsWaitingForConnectivity task: URLSessionTask) {
+        if let netTask = tasks[task] {
+            netTask.state = .waitingForConnectivity
+        }
+    }
+
 }
 
 extension NetURLSessionDelegate: URLSessionDataDelegate {}